requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
ddt==1.7.2
et-xmlfile==1.1.0
exceptiongroup==1.2.0
Faker==18.13.0
Flask==2.2.5
idna==3.6
importlib-metadata==6.7.0
iniconfig==2.0.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.3
jsonpath==0.82.2
MarkupSafe==2.1.5
openpyxl==3.1.2
packaging==24.0
pluggy==1.2.0
ply==3.11
PyMySQL==1.1.0
pytest==7.4.4
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
xlrd==1.2.0
zipp==3.15.0
